Don't reset the list of allowed values when adding new entries
Adding entries from data sources called Kernel:mergeFields() which
used a QSet to merge allowed values. The result is that all the default
values are always added back, and the ordering is not retained.

Move Kernel::mergeFields to Document::mergeFields and retain the
existing list of allowed values, appending new values _if necessary_.

Add a unit test.

BUG: 317905
FIXED-IN: 2.3.8