Correct the name of the image directory when file has multiple periods
The local image directory was using .section('.', 0, 0) and should
be using completeBaseName() so file names with multiple periods would
use the expected directory name

Unit test added to verify that backwards compatability is not broken.
But users with this behavior would not be able to go back to an older
version and have the images work correctly.

BUG: 348088

Conflicts:
	src/images/imagefactory.cpp