Use elided text in Choice combo box to avoid overly wide dialog
Use 40% of screen width as maximum, eliding text in the middle. Change
to using QComboBox instead of KComboBox

The Entry Edit Dialog could still be too wide for the screen if the
label or other UI elements cause a change. I need to investigate the
appropriate use of QSizePolicy

BUG: 362028
FIXED-IN: 2.3.12