Account for a new local directory name when saving to a new location
For some situations, when saving to a new directory location and with
images saved locally, depending on the cache situation, the image might
not be found.

Don't change the local directory until after the images are saved.
Create a new static method for saving outside of the config locations.
Add a unit test.

Conflicts:
	src/document.cpp
	src/images/imagefactory.cpp
	src/tests/documenttest.cpp