Add ugly QPointers to account for potential self-deletion
This is really a case of fixing the symptom and not the root cause,
which appears to be the use of qApp->processEvents() within
TellicoImporter. Rather than not processing events or being smart
enough to properly use event handling, sprinkle some QPointer(this)
checks everywhere that the "this" pointer might end up deleted.

Using QBenchMark in documenttest seems to show this to have worked.