Improve EntryIconModel to act more as proxy for EntryModel for images
Add a new data role to designate the primary image for an entry. Break
out the image request for the model into a separate method so multiple
places can use it.

Also, use an icon for the EntryModel instead of a pixmap which solves
problems of weirdly big images in the tree view.

CCBUG: 386548