Do not clear Fetch Dialog when Search is clicked
Reverting fix for bug 357799. The dialog has a Clear button for a reason
and the workflow in the bug discussion still holds (personally) where I
expect the user to want to perhaps do a single search from mutiple data
sources for the same item and choose the one with the most appropriate
data.

CCBUG: 357799
CCBUG: 386036
FIXED-IN: 3.2