Fix a selection loop that ended up breaking "Filter by Group"
When a group was selected in the GroupView, the entry selection would
trigger with an empty selection. The EntryEditDialog would ultimately call
slotHandleNew() which would clear all the selection, which would then
deselect the group which had been selected in the first place.

To fix it, in the EntryEditDialog, only clear the selection if the
signals are not blocked, which is the case in setContents() with
an empty selection.

BUG: 389931
FIXED-IN: 3.1.2