Clear the selection when the model is reset
Per Qt documentation, reseting a model does not signal a change in the
selection. Connect the primary selection proxy model to the
modelAboutToBeReset signal to catch the selection change.

This was an existing bug in versions prior to 3.1.1 but did not
manifest since a QSet was use to track selected entries and duplicates
were automatically prevented. Once a QList was used to retain the
selection ordering, the bug showed up.

BUG: 391614
FIXED-IN: 3.1.2