Add hack to resize ComboBox in ToolBar when changing collections
The KSelectAction for the field grouping has a ComboBox in the TooBar
but I can't figure out how to force a resize when the collection fields
are updated. Long field titles get truncated and are ugly. Previously,
the hack was to insert an arbitrarily large empty string when the
combobox is first created ot enforce a logn initial size and then hope
for the best.

Now, this is updated to reflow the whole window toolbar layout, as the
ToolBarEditor does. Seems like using a sledgehammer to kill an ant but
it does work.