Avoid crash when loading a different collection type
For some reason, if removeClient/addClient is called in
slotUpdateCollectionToolBar, Tellico crashes if a new collection type
is then loaded. Something to do with the event loop and ref counting
on a shared pointer. The debug trace was unhelpful.

Avoid the crash by resetting the gui through a timer shot