Add method for guessing and normalizing video game platform name
With many video game data sources, the platform names are not always
identical nor equal to Tellico's default spelling or capitalization. Add
an enum for platform types and a slightly better heurestic for guessing
the platform.

Update IGDB and MobyGames to use it. The ones that use XML/XSLT will
need to have more invasive tweaks.