Call QApplication::quit in the MainWindow d'tor
Sometimes, when closing the Tellico main window using the close button
in the window title bar, it would take a while to quit the application.
Strace showed it as a pause while polling eventfd, and I don't know why.
But the way around it is to tell qApp to quit when the main window is
killed.