Avoid saving invalid XML characters while stripping them during loading
Using QXmlDefaultHandler in Tellico 3.3 silently allowed loading of
invalid XML characters, while the switch to QXmlStreamReader in 3.4
resulted in a data file failing to load. In the event of a parsing
error, strip invalid characters and attempt to reload.

In addition, change the DOM implementation policy to NOT save the
invalid characters in the fil to begin with.

BUG: 443845
FIXED-IN: 3.4.2