Avoid unneccesary resets when modifying the collection
Controller::slotCollectionModified() was simply calling
slotCollectionDeleted() and slotCollectionAdded() on the same
collection. Try to be a little more elegant by tracking whether the
modification caused a structural change in the collection, like adding
a new field, updating the allowed valus, or changing table sizes. Then
only update the edit dialog if need be.

Still need to figure out how ot make the detailed view update more
elegant depending on the same change flag.

This might, and hopefully, fixes the crashing bug in the edit dialog
that pops up intermittently when importing an audio CD.