Use a temporary file for large reports to avoid QWebEngine limit
QWebEngine::load is limited to 2MB, which when encoded and otherwise
processed, seems to be as low as 1,200,000 "characters" in a QString.

For large file, write to a temporary file before loading. This does not
take care of any relative links in the HTML, which will need further
work.

BUG: 461391
FIXED-IN: 3.4.5