Update Discogs, MusicBrainze, and iTunes to separate multi-disc tracks
Create new fields for each disc, similar to what the audio file
importer does. Add unit tests. One drawback is that the field will be
visible for all albums in the collection.

MusicBrainz doesn't have a consistent way of representing the disc and
track numbers, so a regular expression is used to try to parse them.

BUG: 479503
FIXED-IN: 4.0