Remove images from local directory when entry values are changed
When the collection is saved, check on which image values were changed
and if any were removed, delete the image from the disk. Since multiple
data files could share the same images in the common application
directory, this behavior only applies to "Store images in directory
relative to data file" option.

BUG: 509244
FIXED-IN: 4.2