Change XML::removeInvalidXml() to check chars as explicitly signed
The loading procedure of the Tellico own XML files has the following
steps:
(1) load the XML file as-is
(2) in case (1) fails because of not well-formed XML:
(2a) use XML::recoverFromBadXMLName() on the data
(2b) in case (2a) did not change anything, use XML::removeInvalidXml()
     on the data
(2c) in case the data changed in either (2a) or (2b), try to load the
     data

XML::recoverFromBadXMLName() (used in 2a) indirectly uses
xmlValidateNCName() from libxml2, and it changed its behaviour:
- before version 2.15: xmlValidateNCName() does not accept any non-ASCII
  name, so XML::recoverFromBadXMLName() results in only ASCII tags,
  and step (2b) is not done
- starting from version 2.15: xmlValidateNCName() accepts Unicode names
  too, so the output of (2a) is unchanged, and thus (2b) is done;
  XML::removeInvalidXml() filters characters by checking them as 'char',
  and on architectures where char is unsigned (e.g. ARM) the existing
  check accepts also Unicode control characters, so the output is not
  changed either

The above is the flow when trying to load the bug418067.xml test case.

Since the step (2b) is a fallback for a fallback, and generally very
much not likely to happen in the real world, then "freeze" the existing
behaviour (i.e. the one happening on architectures with signed char,
such as x86/x86_64).

Drive-by change: make 'c' as const.

(cherry picked from commit 62336cb09439a5f64c2b77dedc9be0741330629b)